test/web-platform-tests/to-upstream/dom/nodes/Node-normalize-2.html

<!DOCTYPE html>
<title>Node.normalize()</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
test(function() {
  var df = document.createDocumentFragment(),
      t1 = document.createTextNode("Hello, world");
  df.appendChild(t1);
  t1.splitText(5);
  var r = document.createRange();
  r.selectNode(t1);
  assert_equals(r.toString(), "Hello");
  df.normalize();
  assert_equals(r.toString(), "Hello");
})

test(function() {
  var df = document.createDocumentFragment(),
      t1 = document.createTextNode("Hello, world");
  df.appendChild(t1);
  var t2 = t1.splitText(7);
  var r = document.createRange();
  r.selectNode(t2);
  assert_equals(r.toString(), "world");
  df.normalize();
  assert_equals(r.toString(), "world");
})
</script>
